requirements.txt
lxml==6.0.1
numpy==2.3.3
pandas==2.3.2
psycopg2-binary==2.9.10
python-dateutil==2.9.0.post0
pytz==2025.2
setuptools==80.9.0
six==1.17.0
tqdm==4.67.1
tzdata==2025.2
wheel==0.45.1
